+++ docs/release-notes.md
@@ -6,12 +6,14 @@
 
 - Add lazy_setup parameter for testclient.
 - `disconnect` and `connect` return if they performed the setup/cleanup.
+- Add backward compatible overwrites of defaults to the TestClient.
 
 ### Fixed
 
 - Restore eager setup behavior in testclient.
 - Fix drop_dabase with postgresql.
 - Fix setup/disconnect hangups.
+- Fix ancient docs deps.
 
 ### Removed